basit hesap makinesi

+++ veri_tipi_donusumleri.py
@@ -1,5 +1,5 @@
 # kullanıcıdan 2 sayı alıp toplama işlemi yapan program
 sayi1 = input("1. sayıyı giriniz :")
 sayi2 = input("2. sayıyı giriniz :")
-print(int(sayi1)+int(sayi2))
+print(float(sayi1)+float(sayi2))
 # basit hesap makinesi
